update configuration to support ios

+++ core/config.rb
@@ -10,17 +10,21 @@ def reset_capybara
     config.exact = false
     config.ignore_hidden_elements = true
     config.visible_text_only = true
+    config.default_wait_time = RSpec.configuration.element_wait_sec
   end
 
-  # Delete some cookies for the site that are hanging around
-  page.driver.browser.manage.delete_cookie('ewokAuth')
-  page.driver.browser.manage.delete_cookie('ewokAuthGuestPass')
-  page.driver.browser.manage.delete_cookie('keepLogin')
-  page.driver.browser.manage.delete_cookie('is_member')
+  # delete cookies and maximize browser for only desktop browsers
+  if (RSpec.configuration.default_browser != Browsers::IPhone) &&
+     (RSpec.configuration.default_browser != Browsers::IPad)
+    # Delete some cookies for the site that are hanging around
+    page.driver.browser.manage.delete_cookie('ewokAuth')
+    page.driver.browser.manage.delete_cookie('ewokAuthGuestPass')
+    page.driver.browser.manage.delete_cookie('keepLogin')
+    page.driver.browser.manage.delete_cookie('is_member')
 
-  # Ensure the browser is maximized to maximize visibility of element
-  page.driver.browser.manage.window.maximize
-
+    # Ensure the browser is maximized to maximize visibility of element
+    page.driver.browser.manage.window.maximize
+  end
 end
 
 # With implicit waits enabled, use of wait_until methods is no longer required. This method will
@@ -29,7 +33,8 @@ SitePrism.configure do |config|
   config.use_implicit_waits = true
 end
 
-#### ENVIRONMENT SETUP ####
+
+#### ANY ENVIRONMENT SETUP FROM SHELL ####
 
 # Use remote web driver, default true
 ENV['OKL_REMOTE_DRIVER'] ||= "true"
@@ -37,22 +42,56 @@ ENV['OKL_REMOTE_DRIVER'] ||= "true"
 # OKL site sub-domain. Default to qa02 if not set
 ENV['OKL_SERVER'] ||= "qa02"
 
-# OKL host ip. Default to grid hub. Or export to 'localhost' on local machine
-ENV['OKL_HOST_IP'] ||= "sfo-qa-grid-hub.corp.onekingslane.biz"
+# OKL default browser, default to firefox
+ENV['OKL_BROWSER'] ||= "Firefox"
 
+browser = Browsers::Firefox
+host_platform = :all
+host_version = ""
+case ENV['OKL_BROWSER'].downcase.chomp
+  when "firefox" then browser = Browsers::Firefox
+  when "chrome" then browser = Browsers::Chrome
+  when "safari" then browser = Browsers::Safari
+  when "ie" then browser = Browsers::InternetExplorer
+  when "android" then browser = Browsers::Android
+  when "iphone"
+    browser = Browsers::IPhone
+    host_platform = :mac
+    host_version = "7.0.3" # ios sdk version
+  when "ipad"
+    browser = Browsers::IPad
+    host_platform = :mac
+    host_version = "7.0.3" # ios sdk version
+end
+
+# if running locally, then put username/password in url, and set host to 'localhost',
+# otherwise set the config options to run on OKL grid
+if ENV['OKL_REMOTE_DRIVER'].eql?("false")
+  default_url = "https://bkitchener:bkitchener123!@#{ENV['OKL_SERVER']}.newokl.com"
+  host_ip = "localhost"
+  remote_driver = false
+else
+  default_url = "https://#{ENV['OKL_SERVER']}.newokl.com"
+  host_ip ="sfo-qa-grid-hub.corp.onekingslane.biz"
+  remote_driver = true
+end
+
+
+
+#### RSPEC CONFIGURATION ###
 RSpec.configure do |config|
   config.include Capybara::DSL
   config.treat_symbols_as_metadata_keys_with_true_values = true
-  config.add_setting :default_browser, :default => Browsers::Firefox
-  config.add_setting :remote_driver, :default => ENV['OKL_REMOTE_DRIVER'].eql?("true") ? true : false
-  config.add_setting :host_ip, :default => ENV['OKL_HOST_IP']
-  config.add_setting :host_platform, :default=>:any
-  config.add_setting :host_version, :default=>""
+  config.add_setting :default_browser, :default => browser
+  config.add_setting :remote_driver, :default => remote_driver
+  config.add_setting :host_ip, :default => host_ip
+  config.add_setting :host_platform, :default=>host_platform
+  config.add_setting :host_version, :default=>host_version
   config.add_setting :host_port, :default => 4444
   config.add_setting :element_wait_sec, :default => 20
   config.add_setting :screenshot_on_failure, :default => true
   config.add_setting :command_logging, :default => true
-  config.add_setting :default_url, :default => "https://bkitchener:bkitchener123!@#{ENV['OKL_SERVER']}.newokl.com"
+  config.add_setting :default_url, :default => default_url
   config.add_setting :ldap_username, :default => "bkitchener"
   config.add_setting :ldap_user_password, :default =>"bkitchener123!"
   config.add_formatter :documentation,'output.txt'
